pyproject.toml
[project]
name = "defi-ds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dlt[duckdb,filesystem]>=1.12.1",
    "jupyter>=1.1.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
]
